+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/write/context/BigQueryIndirectDataSourceWriterContext.java
@@ -108,7 +108,46 @@ public class BigQueryIndirectDataSourceWriterContext implements DataSourceWriter
   }
 
   @Override
+  public void onDataStreamingWriterCommit(long epochId, WriterCommitMessageContext[] messages) {
+    commitMessages(messages, epochId);
+  }
+
+  @Override
+  public void onDataStreamingWriterAbort(long epochId, WriterCommitMessageContext[] messages) {
+    try {
+      logger.warn(
+          "Aborting epoch {} from streaming write {} for table {}",
+          epochId,
+          writeUUID,
+          BigQueryUtil.friendlyTableName(config.getTableId()));
+    } finally {
+      cleanTemporaryGcsPathIfNeeded(epochId);
+    }
+  }
+
+  @Override
   public void commit(WriterCommitMessageContext[] messages) {
+    commitMessages(messages, 0);
+  }
+
+  @Override
+  public void abort(WriterCommitMessageContext[] messages) {
+    try {
+      logger.warn(
+          "Aborting write {} for table {}",
+          writeUUID,
+          BigQueryUtil.friendlyTableName(config.getTableId()));
+    } finally {
+      cleanTemporaryGcsPathIfNeeded(0);
+    }
+  }
+
+  @Override
+  public void setTableInfo(TableInfo tableInfo) {
+    this.tableInfo = Optional.ofNullable(tableInfo);
+  }
+
+  private void commitMessages(WriterCommitMessageContext[] messages, long epochId) {
     logger.info(
         "Data has been successfully written to GCS. Going to load {} files to BigQuery",
         messages.length);
@@ -151,29 +190,12 @@ public class BigQueryIndirectDataSourceWriterContext implements DataSourceWriter
       if (writeDisposition == JobInfo.WriteDisposition.WRITE_TRUNCATE) {
         updateMetadataIfNeeded();
       }
-      logger.info("Data has been successfully loaded to BigQuery");
     } catch (IOException e) {
       throw new UncheckedIOException(e);
     } finally {
-      cleanTemporaryGcsPathIfNeeded();
+      cleanTemporaryGcsPathIfNeeded(epochId);
     }
-  }
-
-  @Override
-  public void abort(WriterCommitMessageContext[] messages) {
-    try {
-      logger.warn(
-          "Aborting write {} for table {}",
-          writeUUID,
-          BigQueryUtil.friendlyTableName(config.getTableId()));
-    } finally {
-      cleanTemporaryGcsPathIfNeeded();
-    }
-  }
-
-  @Override
-  public void setTableInfo(TableInfo tableInfo) {
-    this.tableInfo = Optional.ofNullable(tableInfo);
+    logger.info("Data has been successfully loaded to BigQuery");
   }
 
   void loadDataToBigQuery(List<String> sourceUris, Schema schema) throws IOException {
@@ -202,7 +224,7 @@ public class BigQueryIndirectDataSourceWriterContext implements DataSourceWriter
     BigQueryWriteHelper.updateTableMetadataIfNeeded(sparkSchema, config, bigQueryClient);
   }
 
-  void cleanTemporaryGcsPathIfNeeded() {
-    intermediateDataCleaner.ifPresent(cleaner -> cleaner.deletePath());
+  void cleanTemporaryGcsPathIfNeeded(long epochId) {
+    intermediateDataCleaner.ifPresent(cleaner -> cleaner.deleteEpochPath(epochId));
   }
 }