+++ spark-bigquery-connector-common/src/test/java/com/google/cloud/spark/bigquery/integration/ReadByFormatIntegrationTestBase.java
@@ -15,7 +15,12 @@
  */
 package com.google.cloud.spark.bigquery.integration;
 
+import static com.google.cloud.spark.bigquery.integration.TestConstants.GA4_TABLE;
 import static com.google.common.truth.Truth.assertThat;
+import static org.apache.spark.sql.functions.col;
+import static org.apache.spark.sql.functions.concat;
+import static org.apache.spark.sql.functions.lit;
+import static org.apache.spark.sql.functions.row_number;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assume.assumeThat;
 import static org.junit.Assume.assumeTrue;
@@ -42,6 +47,8 @@ import java.util.stream.Collectors;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Encoders;
 import org.apache.spark.sql.Row;
+import org.apache.spark.sql.expressions.Window;
+import org.apache.spark.sql.expressions.WindowSpec;
 import org.apache.spark.sql.types.DataType;
 import org.apache.spark.sql.types.DataTypes;
 import org.apache.spark.sql.types.StructField;
@@ -261,7 +268,9 @@ public class ReadByFormatIntegrationTestBase extends SparkBigQueryIntegrationTes
                                 "map_field",
                                 LegacySQLTypeName.RECORD,
                                 FieldList.of(
-                                    Field.of("key", LegacySQLTypeName.STRING),
+                                    Field.newBuilder("key", LegacySQLTypeName.STRING)
+                                        .setMode(Field.Mode.REQUIRED)
+                                        .build(),
                                     Field.of("value", LegacySQLTypeName.INTEGER)))
                             .setMode(Field.Mode.REPEATED)
                             .build())))
@@ -314,6 +323,61 @@ public class ReadByFormatIntegrationTestBase extends SparkBigQueryIntegrationTes
     assertThat(schema.apply("foo").dataType()).isEqualTo(timeStampNTZType.get());
     Row row = df.head();
     assertThat(row.get(0)).isEqualTo(dateTime);
+  }
+
+  @Test
+  public void testWindowFunctionPartitionBy() {
+    WindowSpec windowSpec =
+        Window.partitionBy("user_pseudo_id", "event_timestamp", "event_name")
+            .orderBy("event_bundle_sequence_id");
+
+    Dataset<Row> df =
+        spark
+            .read()
+            .format("bigquery")
+            .option("table", GA4_TABLE)
+            .option("readDataFormat", dataFormat)
+            .load()
+            .withColumn("row_num", row_number().over(windowSpec));
+
+    Dataset<Row> selectedDF =
+        df.select("user_pseudo_id", "event_name", "event_timestamp", "row_num");
+
+    assertThat(selectedDF.columns().length).isEqualTo(4);
+    assertThat(
+            Arrays.stream(df.schema().fields())
+                .filter(field -> field.name().equals("row_num"))
+                .count())
+        .isEqualTo(1);
+    assertThat(selectedDF.head().get(3)).isEqualTo(1);
+  }
+
+  @Test
+  public void testWindowFunctionPartitionByWithArray() {
+    assumeTrue("This test only works for AVRO dataformat", dataFormat.equals("AVRO"));
+    WindowSpec windowSpec =
+        Window.partitionBy(concat(col("user_pseudo_id"), col("event_timestamp"), col("event_name")))
+            .orderBy(lit("window_ordering"));
+
+    Dataset<Row> df =
+        spark
+            .read()
+            .format("bigquery")
+            .option("table", GA4_TABLE)
+            .option("readDataFormat", dataFormat)
+            .load()
+            .withColumn("row_num", row_number().over(windowSpec));
+
+    Dataset<Row> selectedDF =
+        df.select("user_pseudo_id", "event_name", "event_timestamp", "event_params", "row_num");
+
+    assertThat(selectedDF.columns().length).isEqualTo(5);
+    assertThat(
+            Arrays.stream(df.schema().fields())
+                .filter(field -> field.name().equals("row_num"))
+                .count())
+        .isEqualTo(1);
+    assertThat(selectedDF.head().get(4)).isEqualTo(1);
   }
 
   static <K, V> Map<K, V> scalaMapToJavaMap(scala.collection.Map<K, V> map) {