+++ spark-bigquery-dsv2/spark-3.5-bigquery/src/test/java/com/google/cloud/spark/bigquery/integration/Spark35ReadByFormatIntegrationTest.java
@@ -15,12 +15,22 @@
  */
 package com.google.cloud.spark.bigquery.integration;
 
+import java.util.Arrays;
+import java.util.Collection;
 import org.apache.spark.sql.types.DataTypes;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 
+@RunWith(Parameterized.class)
 public class Spark35ReadByFormatIntegrationTest extends ReadByFormatIntegrationTestBase {
 
-  public Spark35ReadByFormatIntegrationTest() {
-    super("ARROW", /* userProvidedSchemaAllowed */ false, DataTypes.TimestampNTZType);
+  @Parameterized.Parameters
+  public static Collection<String> data() {
+    return Arrays.asList("AVRO", "ARROW");
+  }
+
+  public Spark35ReadByFormatIntegrationTest(String dataFormat) {
+    super(dataFormat, /* userProvidedSchemaAllowed */ false, DataTypes.TimestampNTZType);
   }
 
   // tests are from the super-class