+++ spark-bigquery-dsv2/spark-4.1-bigquery/src/test/resources/acceptance/spark-dml-create-table-as-select-in-default-dataset.sql
@@ -1,0 +1,7 @@
+CREATE TABLE ${TABLE}_shakespeare OPTIONS (table='bigquery-public-data.samples.shakespeare');
+
+
+CREATE TABLE ${TABLE} AS
+  SELECT word, SUM(word_count) as word_count
+  FROM ${TABLE}_shakespeare
+  WHERE word='spark' GROUP BY word;